refactor(hub): trocear RunsWorkspaceView en hub/runs/ — fase 2 (03/T02)
RunsWorkspaceView baja de 1.895 a 582 líneas extrayendo sin cambios de
comportamiento a app/components/hub/runs/: runPresentation.ts (helpers
puros de parseo de steps, agrupado de workflow, transcript rows,
formatos), RunStepBits.tsx (stat rows, group header, timeline bar, agent
tab bar, step list item, step detail panel), RunRightOverview.tsx y
RunDetailScreen.tsx. La vista mantiene su RunsTab y API pública; imports
muertos podados.
